+++ CHANGELOG.md
@@ -208,6 +208,9 @@ None
   revisions by default (defined by `revsets.op-diff-changes-in`). A new flag,
   `--show-changes-in`, can be used to override this. [#6083](https://github.com/jj-vcs/jj/issues/6083)
 
+* `jj bookmark advance` now includes an `--create-new` flag that will create
+  missing bookmarks.
+
 ### Fixed bugs
 
 * `.gitignore` with UTF-8 BOM can now be parsed correctly.